+++ docs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Next
 
+* Fixed search queries that include the word "and".
+
+# 5.12.0 (2019-01-27)
+
 # 5.11.0 (2019-01-20)
 
 # 5.10.0 (2019-01-13)
@@ -23,6 +27,7 @@
 * Click the plus icon under an equipped item to search for and transfer items in that slot from anywhere in your inventory.
 * Import a CSV file of items with tags and notes to bulk update the tags/notes for all of those items.
 * CSV - Wrap ID in quotes such that its value is not rounded.
+* Added is:reacquirable to show items that can potentially be pulled from your Collection
 
 # 5.8.3 (2019-01-02)